# docs/api/openapi.yaml
openapi: 3.1.0
info:
  description: |
    오픈마켓형 이커머스 REST API. 이 명세는 컨트롤러와 DTO에서 생성된다.

    - 세션·CSRF 절차, 주문 생성의 멱등 키, 오류 응답 형식과 공통 오류 매핑: [contracts.md](https://github.com/EhighG/Ecommerce-API/blob/develop/docs/contracts.md)
    - 역할별 권한과 소유자 검사: [security.md](https://github.com/EhighG/Ecommerce-API/blob/develop/docs/security.md)의 "인가 규칙"
    - 엔드포인트마다 적은 오류는 클라이언트가 분기할 업무 오류다. 오류 응답의 `message`는 예시이고, 같은 `code`라도 더 구체적인 문장이 올 수 있다. 분기는 HTTP 상태와 `code`로만 한다.

    **Try it out(로컬 서버)**
    1. `GET /auth/csrf`를 실행하고 응답의 `token`을 복사한다.
    2. Authorize에서 `csrf`(`X-CSRF-TOKEN` 헤더)에 그 값을 넣는다.
    3. `POST /auth/login`으로 로그인한다. 세션 쿠키는 브라우저가 보낸다.

    로그아웃하거나 세션이 만료되면 1부터 다시 한다.
  title: Ecommerce API
  version: v1
servers:
- description: 로컬 서버
  url: http://localhost:8081/api
...
paths:
...
  /cart-items:
    get:
      operationId: cartItem_findCartItems
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CartItemListRes"
          description: OK
      summary: 내 장바구니
      tags:
      - Cart
    post:
      description: 이미 담긴 상품이면 수량을 더한다. 응답은 장바구니 항목 ID다.
      operationId: cartItem_addItem
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddItemReq"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                type: integer
                format: int64
          description: OK
        "404":
          content:
            application/json:
              examples:
                "2000":
                  summary: 2000 없거나 삭제된 상품
                  value:
                    code: "2000"
                    message: 상품을 찾을 수 없습니다.
              schema:
                $ref: "#/components/schemas/ApiError"
          description: "- `2000`: 없거나 삭제된 상품"
        "409":
          content:
            application/json:
              examples:
                "7001":
                  summary: 7001 같은 상품을 동시에 담아 충돌함. 다시 시도한다
                  value:
                    code: "7001"
                    message: 장바구니 항목이 이미 변경되었습니다. 잠시 후 다시 시도해주세요.
              schema:
                $ref: "#/components/schemas/ApiError"
          description: "- `7001`: 같은 상품을 동시에 담아 충돌함. 다시 시도한다"
      security:
      - csrf: []
      summary: 장바구니에 담기
      tags:
      - Cart
  /cart-items/quantity:
    patch:
      description: 장바구니 항목 ID가 아니라 상품 ID로 항목을 찾는다.
      operationId: cartItem_changeQuantity
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangeQuantityReq"
        required: true
      responses:
        "200":
          description: OK
        "404":
          content:
            application/json:
              examples:
                "7000":
                  summary: 7000 그 상품이 장바구니에 없음
                  value:
                    code: "7000"
                    message: 장바구니 항목을 찾을 수 없습니다.
              schema:
                $ref: "#/components/schemas/ApiError"
          description: "- `7000`: 그 상품이 장바구니에 없음"
      security:
      - csrf: []
      summary: 장바구니 수량 변경
      tags:
      - Cart
  /cart-items/{cartItemId}:
    delete:
      operationId: cartItem_removeItem
      parameters:
      - in: path
        name: cartItemId
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
        "404":
          content:
            application/json:
              examples:
                "7000":
                  summary: 7000 없거나 남의 항목
                  value:
                    code: "7000"
                    message: 장바구니 항목을 찾을 수 없습니다.
              schema:
                $ref: "#/components/schemas/ApiError"
          description: "- `7000`: 없거나 남의 항목"
      security:
      - csrf: []
      summary: 장바구니 항목 삭제
      tags:
      - Cart
  /coupons/events:
    post:
      description: 응답은 이벤트 ID다.
      operationId: coupon_createCouponEvent
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCouponEventReq"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                type: integer
                format: int64
          description: OK
      security:
      - csrf: []
      summary: 쿠폰 이벤트 생성
      tags:
      - Coupon
  /coupons/events/{couponEventId}:
    get:
      operationId: coupon_getCouponEventDetail
      parameters:
      - in: path
        name: couponEventId
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CouponEventDetailRes"
          description: OK
        "404":
          content:
            application/json:
              examples:
                "7500":
                  summary: 7500 이벤트가 없음
                  value:
                    code: "7500"
                    message: 쿠폰 이벤트를 찾을 수 없습니다.
              schema:
                $ref: "#/components/schemas/ApiError"
          description: "- `7500`: 이벤트가 없음"
      summary: 쿠폰 이벤트 조회
      tags:
      - Coupon
  /coupons/events/{couponEventId}/issue:
    post:
      description: 응답의 `couponIssuedId`를 주문할 때 쓴다.
      operationId: coupon_issueCoupon
      parameters:
      - in: path
        name: couponEventId
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IssueCouponRes"
          description: OK
        "400":
          content:
            application/json:
              examples:
                "7503":
                  summary: 7503 발급 기간이 아님
                  value:
                    code: "7503"
                    message: 진행 중인 쿠폰 이벤트가 아닙니다.
                "7506":
                  summary: 7506 이미 이 이벤트의 쿠폰을 받음
                  value:
                    code: "7506"
                    message: 이미 발급받은 쿠폰입니다.
                "7507":
                  summary: 7507 수량이 모두 소진됨
                  value:
                    code: "7507"
                    message: 쿠폰이 모두 소진되었습니다.
              schema:
                $ref: "#/components/schemas/ApiError"
          description: |-
            - `7503`: 발급 기간이 아님
            - `7506`: 이미 이 이벤트의 쿠폰을 받음
            - `7507`: 수량이 모두 소진됨
      security:
      - csrf: []
      summary: 쿠폰 발급(선착순)
      tags:
      - Coupon
...
  /orders:
    post:
      description: "장바구니 항목으로 주문한다. 응답은 주문 ID다. 멱등 키를 만들고 재시도하는 방법은 [contracts.md](https://github.com/EhighG/Ecommerce-API/blob/develop/docs/contracts.md)의\
        \ 주문 생성의 멱등 키 절에 있다."
      operationId: order_order
      parameters:
      - description: "규칙은 [contracts.md](https://github.com/EhighG/Ecommerce-API/blob/develop/docs/contracts.md)의\
          \ 주문 생성의 멱등 키 절에 있다."
        in: header
        name: Idempotency-Key
        required: true
        schema:
          type: string
          example: 3f1e9a2c-7b4d-4c1e-9f0a-2b6c8d1e5a70
          pattern: "^[A-Za-z0-9._:-]{1,128}$"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderReq"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                type: integer
                format: int64
          description: OK
        "400":
          content:
            application/json:
              examples:
                "2004":
                  summary: 2004 상품 삭제와 주문이 동시에 일어남
                  value:
                    code: "2004"
                    message: 삭제된 상품입니다.
                "2501":
                  summary: 2501 재고가 모자람
                  value:
                    code: "2501"
                    message: 주문 가능한 재고가 부족합니다.
                "7501":
                  summary: "7501 쓸 수 없는 쿠폰(이미 사용, 만료)"
                  value:
                    code: "7501"
                    message: 사용할 수 없는 쿠폰입니다.
                "7509":
                  summary: 7509 같은 쿠폰을 두 번 씀
                  value:
                    code: "7509"
                    message: 같은 쿠폰은 중복 사용할 수 없습니다.
                "9001":
                  summary: 9001 한 주문에 같은 장바구니 항목이 두 번 있음
                  value:
                    code: "9001"
                    message: 요청 값이 올바르지 않습니다.
                "9100":
                  summary: 9100 `Idempotency-Key` 헤더가 없음
                  value:
                    code: "9100"
                    message: Idempotency-Key 헤더가 없습니다.
                "9101":
                  summary: 9101 `Idempotency-Key` 형식이 틀림
                  value:
                    code: "9101"
                    message: Idempotency-Key 형식이 올바르지 않습니다.
              schema:
                $ref: "#/components/schemas/ApiError"
          description: |-
            - `9100`: `Idempotency-Key` 헤더가 없음
            - `9101`: `Idempotency-Key` 형식이 틀림
            - `9001`: 한 주문에 같은 장바구니 항목이 두 번 있음
            - `9001`: 할인액이 0원이 되는 쿠폰을 씀. 메시지는 경우마다 다르다
            - `2004`: 상품 삭제와 주문이 동시에 일어남
            - `7509`: 같은 쿠폰을 두 번 씀
            - `7501`: 쓸 수 없는 쿠폰(이미 사용, 만료)
            - `2501`: 재고가 모자람
        "404":
          content:
            application/json:
              examples:
                "7000":
                  summary: 7000 장바구니 항목이 없거나 남의 항목. 상품을 삭제하면 장바구니 항목도 지워지므로 삭제된 상품도
                    대개 이 코드다
                  value:
                    code: "7000"
                    message: 장바구니 항목을 찾을 수 없습니다.
                "7510":
                  summary: 7510 쿠폰이 없거나 남의 쿠폰
                  value:
                    code: "7510"
                    message: 해당 쿠폰을 찾을 수 없습니다.
              schema:
                $ref: "#/components/schemas/ApiError"
          description: |-
            - `7000`: 장바구니 항목이 없거나 남의 항목. 상품을 삭제하면 장바구니 항목도 지워지므로 삭제된 상품도 대개 이 코드다
            - `7510`: 쿠폰이 없거나 남의 쿠폰
        "409":
          content:
            application/json:
              examples:
                "9102":
                  summary: 9102 같은 키로 다른 내용을 보냄
                  value:
                    code: "9102"
                    message: 같은 Idempotency-Key로 다른 요청을 보낼 수 없습니다.
                "9103":
                  summary: 9103 같은 키의 요청이 처리 중. 잠시 후 같은 키로 다시 보낸다
                  value:
                    code: "9103"
                    message: 동일한 요청이 이미 처리 중입니다.
              schema:
                $ref: "#/components/schemas/ApiError"
          description: |-
            - `9102`: 같은 키로 다른 내용을 보냄
            - `9103`: 같은 키의 요청이 처리 중. 잠시 후 같은 키로 다시 보낸다
      security:
      - csrf: []
      summary: 주문 생성
      tags:
      - Order
...
  /orders/{orderId}:
    get:
      operationId: order_getMyOrderDetail
      parameters:
      - in: path
        name: orderId
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderDetailRes"
          description: OK
        "404":
          content:
            application/json:
              examples:
                "3000":
                  summary: 3000 주문이 없음
                  value:
                    code: "3000"
                    message: 주문을 찾을 수 없습니다.
                "3001":
                  summary: 3001 남의 주문
                  value:
                    code: "3001"
                    message: 주문을 찾을 수 없습니다.
              schema:
                $ref: "#/components/schemas/ApiError"
          description: |-
            - `3000`: 주문이 없음
            - `3001`: 남의 주문
      summary: 주문 상세
      tags:
      - Order
...
components:
  schemas:
...
    AddItemReq:
      type: object
      properties:
        productId:
          type: integer
          format: int64
        quantity:
          type: integer
          format: int32
          minimum: 1
      required:
      - productId
      - quantity
    ApiError:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
...
    CartItemListRes:
      type: object
      properties:
        id:
          type: integer
          format: int64
        linePrice:
          type: integer
          format: int64
        product:
          $ref: "#/components/schemas/ProductListRes"
        quantity:
          type: integer
          format: int32
...
    ChangeQuantityReq:
      type: object
      properties:
        productId:
          type: integer
          format: int64
        quantity:
          type: integer
          format: int32
          minimum: 1
      required:
      - productId
      - quantity
...
    CouponEventDetailRes:
      type: object
      properties:
        couponEventId:
          type: integer
          format: int64
        discountValue:
          type: integer
          format: int64
        endAt:
          type: string
          format: date-time
        initialQuantity:
          type: integer
          format: int32
        maxDiscountAmount:
          type: integer
          format: int64
        name:
          type: string
        open:
          type: boolean
        remainingQuantity:
          type: integer
          format: int32
          description: "남은 수량. 이벤트가 Redis에 올라가 있지 않으면 `null`이다. 발급 시작 10분 전보다 이를 때\
            , 종료 10분 뒤부터, Redis를 잃고 다시 적재하기 전이 그렇다. 적재는 최대 30초 늦을 수 있다."
        startAt:
          type: string
          format: date-time
        type:
          type: string
          enum:
          - FIXED_AMOUNT
          - PERCENT
        validSeconds:
          type: integer
          format: int64
    CreateCouponEventReq:
      type: object
      properties:
        discountValue:
          type: integer
          format: int64
          description: "`FIXED_AMOUNT`이면 할인 금액(원), `PERCENT`이면 할인율(%)"
        durationUnit:
          type: string
          enum:
          - MINUTES
          - HOURS
          - DAYS
        endAt:
          type: string
          description: "발급 종료 시각. `yyyy-MM-dd HH:mm:ss`, `timezone` 기준"
          example: 2026-10-08 23:59:59
          minLength: 1
        initialQuantity:
          type: integer
          format: int32
        maxDiscountAmount:
          type: integer
          format: int64
          description: 할인액 상한(원)
        name:
          type: string
          minLength: 1
        startAt:
          type: string
          description: "발급 시작 시각. `yyyy-MM-dd HH:mm:ss`, `timezone` 기준"
          example: 2026-10-01 10:00:00
          minLength: 1
        timezone:
          type: string
          description: IANA 시간대 ID. 비우면 `Asia/Seoul`
          example: Asia/Seoul
        type:
          type: string
          enum:
          - FIXED_AMOUNT
          - PERCENT
        validDurationAmount:
          type: integer
          format: int64
          description: 발급 시점부터 쿠폰을 쓸 수 있는 기간. 단위는 `durationUnit`
      required:
      - durationUnit
      - endAt
      - name
      - startAt
      - type
...
    IssueCouponRes:
      type: object
      properties:
        couponEventId:
          type: integer
          format: int64
        couponIssuedId:
          type: integer
          format: int64
        expiresAt:
          type: string
          format: date-time
        issuedAt:
          type: string
          format: date-time
...
    OrderDetailRes:
      type: object
      properties:
        currentTotalPrice:
          type: integer
          format: int64
        itemList:
          type: array
          items:
            $ref: "#/components/schemas/OrderItemListRes"
        orderId:
          type: integer
          format: int64
        orderedAt:
          type: string
          format: date-time
        totalPrice:
          type: integer
          format: int64
...
    OrderItemListRes:
      type: object
      properties:
        deliveredAt:
          type: string
          format: date-time
        finalLinePrice:
          type: integer
          format: int64
        linePrice:
          type: integer
          format: int64
        orderItemId:
          type: integer
          format: int64
        product:
          $ref: "#/components/schemas/ProductSummary"
        quantity:
          type: integer
          format: int32
        status:
          type: string
          enum:
          - ORDERED
          - SHIPPED
          - DELIVERED
          - PURCHASE_CONFIRMED
          - CANCELED
        usedCoupon:
          $ref: "#/components/schemas/UsedCouponSummary"
    OrderItemReq:
      type: object
      properties:
        cartItemId:
          type: integer
          format: int64
        couponIssuedId:
          type: integer
          format: int64
          description: 쿠폰 발급 응답의 `couponIssuedId`. 쿠폰을 쓰지 않으면 비운다
        orderQuantity:
          type: integer
          format: int32
      required:
      - cartItemId
      - orderQuantity
...
    OrderReq:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItemReq"
          minItems: 1
      required:
      - items
...
    ProductListRes:
      type: object
      properties:
        avgRating:
          type: number
          format: double
        categoryName:
          type: string
        id:
          type: integer
          format: int64
        inventoryQuantity:
          type: integer
          format: int32
        name:
          type: string
        seller:
          $ref: "#/components/schemas/UserSummary"
        thumbnailImageUrl:
          type: string
        unitPrice:
          type: integer
          format: int64
...
    ProductSummary:
      type: object
      properties:
        name:
          type: string
        productId:
          type: integer
          format: int64
        thumbnailImageUrl:
          type: string
...
    UsedCouponSummary:
      type: object
      properties:
        couponIssuedId:
          type: integer
          format: int64
        discountAmount:
          type: integer
          format: int64
        discountValue:
          type: integer
          format: int64
        maxDiscountAmount:
          type: integer
          format: int64
        name:
          type: string
        type:
          type: string
          enum:
          - FIXED_AMOUNT
          - PERCENT
        usedAt:
          type: string
          format: date-time
...
    UserSummary:
      type: object
      properties:
        nickname:
          type: string
        userId:
          type: integer
          format: int64
...
  securitySchemes:
    csrf:
      description: "`GET /auth/csrf` 응답의 `token`. POST, PATCH, DELETE 요청에 필요하다."
      in: header
      name: X-CSRF-TOKEN
      type: apiKey